resolves #24 Create Real-time community Function and Page (#25)
* Spring init

* #1 Jwt and Security config

* #1 User Entity, Dto, Service, Repository

* #1 User Controller init

* java version update

* #1 Signup, Signin, Response Dto

* #1 Update User Entity, Repository

* #1 Signup, Signin Controller

* #1 application.yaml config h2 database

* #1 fix CustomUserdetails

* #3 Change project folder name

* #3 Config CORS mapping

* #3 Create react app

* fix files to commit

* delete error folder

* #3 Set SigninResponse return with username

* #3 frontend layouts

* #3 frontend signup, signin design and link to backend

* #3 회원명 중복 검사 로직

* #3 주석 수정 및 약간의 오류 수정

* #3 회원명 중복 검사 api 연동

* #3 내비게이션 로고 크기 및 클릭 범위 수정

* #5 Posts Entity

* #5 Posts Controller and Dto

* #5 Posts Service and Repository

* #7 Posts CRUD Page

* #7 주석 수정

* #7 Update Posts Dto

* #7 검색 및 페이지네이션 기능 구현

* #7 검색과 페이지네이션과 Dto 업데이트 및 수정

* #7 Floara WYSIWYG 기본 설정 및 약간의 오류 수정

* #7 Post 전체 가져오기 및 검색 기능 완성

* #7 Floara WYSIWYG을 사용한 Post 생성 기능 완성

* 리베이스 진행 용 커밋

* resolves #5 Create post backend function (#6)

* Spring init

* #1 Jwt and Security config

* #1 User Entity, Dto, Service, Repository

* #1 User Controller init

* java version update

* #1 Signup, Signin, Response Dto

* #1 Update User Entity, Repository

* #1 Signup, Signin Controller

* #1 application.yaml config h2 database

* #1 fix CustomUserdetails

* #3 Change project folder name

* #3 Config CORS mapping

* #3 Create react app

* fix files to commit

* delete error folder

* #3 Set SigninResponse return with username

* #3 frontend layouts

* #3 frontend signup, signin design and link to backend

* #3 회원명 중복 검사 로직

* #3 주석 수정 및 약간의 오류 수정

* #3 회원명 중복 검사 api 연동

* #3 내비게이션 로고 크기 및 클릭 범위 수정

---------


* Revert "resolves #5 Create post backend function (#6)" (#8)

This reverts commit 389d164952af0a267635cef4181353c048d57ce8.

* 병합용 커밋

* #7 Json Web Token 전송 방식을 Cookie로 변경(보안 향상)

* #7 database ddl-auto 설정 변경

* #7 Post 관리 페이지 디자인 및 기능 추가, 수정

* #7 Post 삭제 및 좋아요 기능 구현

* #7 주석 삭제 및 추가

* #11 모바일 화면에서도 아이콘 보이게 수정

* #11 Comment, CommentLikes Entity

* #11 파일 이름 변경 및 위치 변경

* #11 Comment Controller, Dto

* #11 Comment Service, Repository

* #13 Comment CRUD page and link with backend function

* #14 Create myPage function + 내 게시판, 본 게시글 저장, 수정 내역 저장

* #16 Create my page and etc design link with backend function

* 병합용 커밋

* #18 My page, My Favorite, My Actives

* 병합용 커밋

* #18 Create Favorite and My Actives Function

* #18 Create Favorite and My Actives Pages and delete alert

* #19 Alert Function 일부 및 질문 게시판에서 댓글 채택 기능

* #19 Alert Page 및 질문 게시글에서의 채택 버튼 표시 및 이미 채택된 댓글 다르게 표시

* #19 공통된 색상 및 함수를 독립된 하나의 파일로 만들어 관리

* #19 Posts에 maxUserNumber(최대 인원수), currentUserNumber(현재 인원수) 추가

* #19 모임 신청 로직

* #19 알림 Entity, Dto, Controller, Service, Repository

* #19 신청 알림 승인 및 거절 로직

* #19 알림 읽음 상태 로직 비활성화(오류 발생: 추가 후에 바로 삭제하면 안됨)

* #19 알림을 읽지 않았을 때 녹색 표시 비활성화(오류)

* #19 모임 신청 상태에 따라 모임 신청란 다르게 표시

* #19 회원가입 이메일 검증 Page

* #19 읽은 알림 구분을 기능 및 UserViewed -> PostsViewed로 이름 변경

* #19 useLocation을 이용해서 매 랜더링마다 새로운 알림이 있는지 검사 및 표시, 읽지 않은 알림을 NEW_COLOR로 표시

* #19 UserViewedRepository.java -> PostsViewedRepository.java 이름 변경

* #22 Email Validate page

* #22 이메일에 코드 전송을 할 때 이메일 중복 검사 실행

* #22 git ignore

* #22 이메일 서버(Gmail SMTP) 및 Redis 연결 정보 설정

* #22 인증 코드 생성 클래스

* #22 인증코드 전송 및 검사 서비스

* #22 약간의 오류 수정

* #22 이메일 인증 컨트롤러 및 이메일 의존성 추가

* #22 이메일 인증 api 연결

* #22 이메일 인증 후 권한 TEMP -> 비밀번호 재설정 로직에 TEMP 권한을 가진 사용자만 접근할 수 있도록 설정

* #22 비밀번호 재설정 페이지, axios 최대 대기 가능 시간 증가(이메일 보내는게 오래걸림), 디자인 수정

* #22 비밀번호 재설정 페이지

* #24 Implement Web Socket

* #24 백엔드 웹소켓 설정 (기능 api 미구현)

* #24 백엔드 웹소켓 설정 및 기능 일부 생성 (메모리 -> 데이터베이스)

* #24 프론트엔드 리스트 필드 크기 조절 및 폰트 notoSansKr로 변경, 커뮤니티 리스트 페이지 및 커뮤니티 대화 페이지 생성

* #24 약간의 오류 수정 및 정리

* #24 Real-time community function, 게시글 삭제 오류 있음

* #24 Real-time community Page

---------

+++ frontend/src/components/user/MyPage.js
@@ -246,6 +246,7 @@ const MyPage = () => {
   };
 
   const handleCancellation = async () => {
+    if(window.confirm("정말 회원을 탈퇴하시겠습니까?")) {
         apiClient.delete("/user/my/cancellation").then(res => {
             logout()
         }).catch(err => {
@@ -253,7 +254,7 @@ const MyPage = () => {
             const error = err.response.data.message || '처리되지 않은 예외'
             alert(error)
         })
-
+      }
   }
 
   return (
